rename `datacore-view-page` -> `view-page`
